feat(ui): moves add inventory details to steps dir
